[replacelist] stop needlessly writing timestamps...
in order to stop causing files to be updated when they don't need to be
(because we're running again from the same machine we did last time, so
the height and width of things will be the same), check that we don't
have our machine's height and width first before acting.

Also made this functionality into a generic script for the simple case
where we have height and width and some prefix for the variable name,
its only active for one emu but can now be extended to many more. I
don't like the way we have to specify the key plus separator but the
alternative appears horrid