[replacelist] restore previous generic fn algo...
its faster, a lot less elegant. The new algorithm didnt
short-circuit the initial search for wrong props, it had to search
through all of width/height/res

and then it compared each line of the file again to see if it had
changed before saving, and you have to check object equality not
identity as its a functional pipe. This was making it about 7-10 seconds
slower. The new algo may still be useful for the other system props
hence keeping the commit