[nas sync] attempt to schedule config and screen syncs...
plan was to sync screenshots as a standard realtime sync ie: when the
directory contents change, but to sync used emu files only when qp is
open but commonly used emus are not open. However putting these syncs in
the frontend runner rather negates the latter strategy and anyway it
seems realtime sync will always fire up on first run. So now in
particular the second sync (used emu config) will run on every file
change of emu drive, which is not appropriate