[nas sync] prototype powershell-only sync fn
started here (and it works but isn't inline) a powershell version of the
sync functionality in the runner.bat, in the expectation that it would
be easier to mirror the functionality so far and then apply the check
whether QuickPlay or used emulators were open when deciding whether to
cascade syncs. Got as far as making the Check-Running fn, which surely
would be solvable in short time, but realised that this whole strategy
is only appealing because of the technical challenge and won't actually
be sensible in operation: we really only want to sync files that a
machine saves (or manual game saves) in a sync when QP opens. Other
types of syncing are a manual process. So i'll back out of this: will we
really ever be happy touching most all files in the emu folder on two
machines twice when using our frontend to play games? Of course not....