[nas sync] only one sync job runs on qp open/close...
the goal is to tweak the job for machine-controlled file updates.
Have been able to retain a nice feature of the previous strategy: if the
sync is still running when QP exits, stop any syncing. If not: sync
again.

However there is now no prospect of stopping the sync if an emu is
launched, but do note the previous plan was just to not begin a further
stage if emu launched. In order to mitigate issues when emu is launched,
start the single new sync job in LOW system priority (which some sources
say has to be launched as admin but seems to work for me in this context
without that)