[nas sync] externalise sycing code from qp runner...
...with intention to next integrate this code into switch-to-tv script
an interesting thing to note here is the pushd and popd that's needed in
the runner, took some care that the script could still one day be run
from somewhere other than P: if needed.

also integrated the logic of killing the ffs window and not beginning a
sync if ffs is currently running, this changes the behaviour to do this
both on now entering as well as exiting frontend. (previously ran only
on exit)