[nas sync] save windows game files to emu drive...
this functionality is designed to remove the need for the peer-peer
administrator realtime sync entirely (ie: the contents of the
RIVER_Emu_NoNAS_Sync folder here, which hopefully is now defunct)

The function is a little subtle, we realime sync for changes in my
documents and my games folders, and put those items in my emulator
drive (so its a sync from a drive to itself in effect). We then will
capture the effect of that sync whenever we run the games sync when the
frontend is open (which is a nice separation of concerns - we're
thinking about syncing games all in one place)

the only thing i don't like about this is that it doesn't REALLY need to
be realtime synced, the only reason it is is the 'same-drive' sync isn't
really about emulator games and so syncing when quickplay is opened
isn't quite right, but is that really worth another sync runner process?