[gamebase export] standaardise reg export process...
we had one process loading the reg from the gamebase dir,
a deactivated process saving it there, and another process saving the
same reg file to this repo. Not only do we standardise this, we
re-enable the save in the runner scripts (still pretty un-DRY), we keep
an archive of previus reg files (to safeguard an accident really), and
we make the reg location the the archive be the gamebase dir (this repo
is a code. not a data, store).

Note there's another process in the win-autoinstall script that loads
the gamebase reg at install time, but its location remamins the reg in
the gamebase dir